Add max bio size param
We observed that when the number of connections is very large
on a service, the machines started running out of memory.
On memory analysis the major consumer of memory was found
to be the bio allocation.

The default max bio size is 17K. This allows the parameter to be
tunable.
We've found this extremely useful for scaling large services. The
CPU usage goes up a bit when this is reduced so everyone find their
own sweet spot.